Refactor : 기존 Card 컴포넌트 삭제, CardItem 컴포넌트명을 Card로 변경
