+++ src/components/Card/Card.js
@@ -1,31 +1,61 @@
 import styled from "@emotion/styled";
 
-import CardItem from "../CardItem";
+import Item from "../style/card";
 
-const Card = ({ card }) => {
-  const { name, numbers, owner, expiryDate, nickname } = card;
-
-  const data = { name, numbers, owner, expiryDate };
+const CardItem = ({ data }) => {
+  const {
+    name,
+    numbers: [firstField, secondField],
+    owner,
+    expiryDate,
+  } = data;
 
   return (
-    <Wrap>
-      <CardItem data={data} />
-      <Nickname>{nickname}</Nickname>
-    </Wrap>
+    <Item>
+      <Top>{name}</Top>
+      <Middle>
+        <span>{firstField}</span>
+        <span>{secondField}</span>
+        <span>●●●●</span>
+        <span>●●●●</span>
+      </Middle>
+      <Bottom>
+        <span>{owner}</span>
+        <span>{expiryDate}</span>
+      </Bottom>
+    </Item>
   );
 };
 
-const Wrap = styled.li`
-  & + & {
-    margin-top: 20px;
+const Top = styled.p`
+  font-size: 10px;
+  &:after {
+    display: block;
+    margin: 20px 0 14px;
+    width: 40px;
+    height: 25px;
+    border-radius: 5px;
+    background: #cbba64;
+    content: "";
   }
 `;
 
-const Nickname = styled.h2`
-  margin-top: 10px;
-  font-size: 14px;
-  text-align: center;
-  color: #575757;
+const Middle = styled.p`
+  display: flex;
+  justify-content: space-between;
+  font-size: 12px;
+
+  span {
+    vertical-align: middle;
+  }
 `;
 
-export default Card;
+const Bottom = styled.p`
+  display: flex;
+  margin-top: auto;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 12px;
+`;
+
+export default CardItem;
