--- src/components/CardItem/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CardItem";
